updated posterior description
